Embed CM9.3 manual and update faiss index creation script
This commit embeds the CM9.3 manual in the chatbot and updates the create_faiss_index.py script to save the faiss index locally. The manual used is from CM9.3, but it requires additional information to be supplemented. The faiss index creation script can be executed using the command "python create_faiss_index.py --json_path [JSON file path] --output_path [faiss index save path]". The default value for output_path is "db/". It is recommended to use the default value.

This commit message follows the established convention of using a prefix to indicate the type of change (feat for feature), followed by a concise summary of the changes made.