feat: Update create_faiss_index.py to use conda environment for setup
This commit updates the README.md file to include instructions for creating the conda environment using the command "conda env create -f environment.yml". It also adds a section explaining how to activate the environment using the command "conda activate [environment name]".

This commit message follows the established convention of using a prefix to indicate the type of change (feat for feature), followed by a concise summary of the changes made.